% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/gs_get_coordinates.R
\name{gs_get_coordinates}
\alias{gs_get_coordinates}
\title{Retrieve coordinates and information of an address}
\usage{
gs_get_coordinates(
  query,
  index = "address",
  limit = 10,
  postcode = NULL,
  type = NULL,
  city = NULL,
  category = NULL,
  longitude = NULL,
  latitude = NULL
)
}
\arguments{
\item{query}{a \code{character} of length 1. The address or point of interest to
search for. See example.}

\item{index}{a \code{character} of length 1. The type of location. Possible
values: \code{address} and \code{poi}.}

\item{limit}{a \code{numeric} of length 1. The maximum number of results to
return (between 1 and 20).}

\item{postcode}{a \code{numeric} of length 1 (optional). Can be used to filter
results based on the postal code.}

\item{type}{a \code{character} of length 1 (optional). Can be used to filter
results based on the type of address. Possible values : \code{housenumber},
\code{street}, \code{locality}, \code{municipality}. This argument only works if
\code{index = 'address'}.}

\item{city}{a \code{character} of length 1 (optional). Can be used to filter
results based on a city name.}

\item{category}{a \code{character} of length 1 (optional). Can be used to filter
results based on the type of POI. Run \code{gs_get_poi_categories()} to get
accepted values. This argument only works if \code{index = 'poi'}.}

\item{longitude}{a \code{numeric} of length 1 (optional). Can be used to filter
results based on the longitude of a known near point.}

\item{latitude}{a \code{numeric} of length 1 (optional). Can be used to filter
results based on the latitude of a known near point.}
}
\value{
A \code{data.frame} with the following columns:
\itemize{
\item \code{query}: the user query
\item \code{longitude}: the longitude of the location matching the query
\item \code{latitude}: the latitude of the location matching the query
\item \code{label}: the label of the location (combination of name, postal & city)
\item \code{name}: the name of the location in the BD TOPO® system
\item \code{toponym}: the toponym of the point of interest
\item \code{housenumber}: the housenumber of the address
\item \code{street}: the street name of the address
\item \code{city}: the city of the address
\item \code{postcode}: the postal code of the city
\item \code{citycode}: the COG (INSEE) code of the city
\item \code{context}: a combination of the department and region
\item \code{category}: the type of point of interest
\item \code{score}: the relevance of the match (used to sort results)
}

Depending of the value of \code{index} (\code{address} of \code{poi}) some columns can be
\code{NA}.
}
\description{
Queries the Geoservices (IGN) API using the \strong{/geocode} endpoint to
extract the coordinates and administrative information of a French address
or point of interest (POI) as cityhall, museum, etc. This service uses the
IGN BD TOPO® database.

The API documentation is available \href{https://geoservices.ign.fr/documentation/services/services-geoplateforme/geocodage}{here}.
}
\examples{
\dontrun{
# Search by full postal address ----

gs_get_coordinates(query = "1 Place Georges Freche 34000 Montpellier", 
                   index = "address")


# Search by full postal address (better results) ----

gs_get_coordinates(query = "1 Place Georges Freche", 
                   index = "address", 
                   type  = "housenumber",
                   city  = "Montpellier")


# Search by point of interest (building) ----

gs_get_coordinates(query    = "Mairie de Montpellier", 
                   limit    = 10, 
                   index    = "poi", 
                   category = "mairie",
                   citycode = 34000)


# Search by point of interest (lake) ----

dat <- gs_get_coordinates(query    = "Lac d'Annecy", 
                          index    = "poi", 
                          category = "lac")


# Map coordinates ----
xy <- sf::st_as_sf(dat, coords = 2:3, crs = sf::st_crs(4326))

mapview::mapview(xy)
} 
}
