% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/gs_get_elevation.R
\name{gs_get_elevation}
\alias{gs_get_elevation}
\title{Retrieve the elevation of French locations}
\usage{
gs_get_elevation(longitude, latitude)
}
\arguments{
\item{longitude}{a \code{numeric} vector. The longitude of one or several French
locations defined in the WGS84 system.}

\item{latitude}{a \code{numeric} vector. The latitude of one or several French
locations defined in the WGS84 system.}
}
\value{
A \code{data.frame} with three columns: \code{longitude}, \code{latitude}, and
\code{elevation}. The number of rows matches the number of provided locations.
}
\description{
Queries the Geoservices (IGN) API using the \strong{/altimetrie} endpoint to
extract the elevation of one or many French (mainland) locations. This
service uses the IGN RGE ALTI® database.

The API documentation is available \href{https://data.geopf.fr/altimetrie/swagger-ui/index.html#/Elevation/get_elevations_1_0_calcul_alti_rest_elevation__output_format__get}{here}.
}
\examples{
\dontrun{
# Set Florac-Trois-Rivieres, Lozere coordinates ----

florac <- data.frame(longitude  =  3.59608777,
                     latitude   = 44.32690171)


# Get Florac-Trois-Rivieres, Lozere elevation ----

ele <- gs_get_elevation(longitude = florac$"longitude", 
                        latitude  = florac$"latitude")
}
}
