% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/gs_get_geocode.R
\name{gs_get_geocode}
\alias{gs_get_geocode}
\title{Retrieve an address or a point of interest from coordinates}
\usage{
gs_get_geocode(
  longitude,
  latitude,
  index = "address",
  limit = 10,
  postcode = NULL,
  type = NULL,
  city = NULL,
  category = NULL
)
}
\arguments{
\item{longitude}{a \code{numeric} of length 1. The longitude of the location
(in WGS84).}

\item{latitude}{a \code{numeric} of length 1. The latitude of the location
(in WGS84).}

\item{index}{a \code{character} of length 1. The type of location to return.
Possible values: \code{address} and \code{poi}.}

\item{limit}{a \code{numeric} of length 1. The maximum number of results to
return (between 1 and 20).}

\item{postcode}{a \code{numeric} of length 1 (optional). Can be used to filter
results based on the postal code.}

\item{type}{a \code{character} of length 1 (optional). Can be used to filter
results based on the type of address. Possible values : \code{housenumber},
\code{street}, \code{locality}, \code{municipality}. This argument only works if
\code{index = 'address'}.}

\item{city}{a \code{character} of length 1 (optional). Can be used to filter
results based on a city name.}

\item{category}{a \code{character} of length 1 (optional). Can be used to filter
results based on the type of POI. Run \code{gs_get_poi_categories()} to get
accepted values. This argument only works if \code{index = 'poi'}.}
}
\value{
A \code{data.frame} with the following columns:
\itemize{
\item \code{longitude}: the longitude of the matched location
\item \code{latitude}: the latitude of the matched location
\item \code{label}: the label of the location (combination of name, postal & city)
\item \code{name}: the name of the location in the BD TOPO® system
\item \code{toponym}: the toponym of the point of interest
\item \code{housenumber}: the housenumber of the address
\item \code{street}: the street name of the address
\item \code{city}: the city of the address
\item \code{postcode}: the postal code of the city
\item \code{citycode}: the COG (INSEE) code of the city
\item \code{context}: a combination of the department and region
\item \code{category}: the type of point of interest
\item \code{distance}: the distance to the provided coordinates (in meters)
\item \code{oldcity}: the old name of the city
\item \code{oldcitycode}: the INSEE code of the old name
\item \code{score}: the relevance of the match (used to sort results)
}

Depending of the value of \code{index} (\code{address} of \code{poi}) some columns can be
\code{NA}.
}
\description{
Queries the Geoservices (IGN) API using the \strong{/geocode} endpoint to
retrieve administrative information of a French address or point of interest
(POI) as cityhall, museum, etc from geographical coordinates. This service
uses the IGN BD TOPO® database.

The API documentation is available \href{https://geoservices.ign.fr/documentation/services/services-geoplateforme/geocodage}{here}.
}
\examples{
\dontrun{
# Search for a postal address ----

gs_get_geocode(longitude = 3.896549, 
               latitude  = 43.59884, 
               index     = "address")


# Search for a point of interest (building) ----

gs_get_geocode(longitude = 3.896549, 
               latitude  = 43.59884, 
               index     = "poi")


# Search for a point of interest (lake) ----

gs_get_geocode(longitude = 6.186529, 
               latitude  = 45.84561, 
               index     = "poi")
} 
}
