% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/gs_get_isochrone.R
\name{gs_get_isochrone}
\alias{gs_get_isochrone}
\title{Calculate an isochrone around a French location}
\usage{
gs_get_isochrone(
  longitude,
  latitude,
  cost_value,
  profile = "car",
  direction = "departure",
  time_unit = "hour"
)
}
\arguments{
\item{longitude}{a \code{numeric} of length 1. The longitude of the location
defined in the WGS84 system. The location must belongs to Mainland France.}

\item{latitude}{a \code{numeric} of length 1. The latitude of the location
defined in the WGS84 system. The location must belongs to Mainland France.}

\item{cost_value}{a \code{numeric} of length 1. The time used to calculate the
isochrone. Unit is specified by the argument \code{time_unit}. Must be strictly
positive.}

\item{profile}{a \code{character} of length 1. The means of transport used to
calculate the isochrone. Possible values: \code{car} and \code{pedestrian}.}

\item{direction}{a \code{character} of length 1. The course direction.
If \code{direction = "departure"}, the isochrone will be calculated from the
starting point (defined by \code{longitude} and \code{latitude} arguments) and will
provide the potential arrival points.
If \code{direction = "arrival"}, the isochrone will provide the potential
departure points to the arrival point (defined by \code{longitude} and
\code{latitude} arguments).}

\item{time_unit}{a \code{character} of length 1. The time unit of the argument
\code{cost_value}. Possible values: \code{hour}, \code{minute} and \code{second}.}
}
\value{
The isochrone as a \verb{sf POLYGON} object.
}
\description{
A short description...
}
\examples{
\dontrun{
# Set Florac-Trois-Rivieres, Lozere coordinates ----

florac <- data.frame(longitude  =  3.59608777559441,
                     latitude   = 44.32690171172478)


# Isochrone of one hour around Florac-Trois-Rivieres, Lozere (by car) ----

iso_car <- gs_get_isochrone(longitude  = florac$"longitude", 
                            latitude   = florac$"latitude", 
                            cost_value = 60,
                            time_unit  = "minute",
                            profile    = "car")


# Isochrone of one hour around Florac-Trois-Rivieres, Lozere (on foot) ----

iso_foot <- gs_get_isochrone(longitude  = florac$"longitude", 
                             latitude   = florac$"latitude", 
                             cost_value = 60,
                             time_unit  = "minute",
                             profile    = "pedestrian")


# Convert point to sf object ----

florac <- sf::st_as_sf(florac, coords = 1:2, crs = sf::st_crs(4326))


# Make map ----

library(ggplot2)

ggplot() +
  geom_sf(data = iso_car,  fill = "red", col = "white") +
  geom_sf(data = iso_foot, fill = "darkgreen", col = "white") +
  geom_sf(data = florac, col = "white", shape = 19) +
  theme_bw()
}
}
