% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/gs_get_poi_categories.R
\name{gs_get_poi_categories}
\alias{gs_get_poi_categories}
\title{Get point of interest categories}
\usage{
gs_get_poi_categories()
}
\value{
A \code{data.frame} with two columns: \code{type} and \code{category} name.
}
\description{
Queries the Geoservices (IGN) API using the \strong{/geocode} endpoint to
list accepted categories of points of interest (POI) such as cityhall,
museum, lake, etc. This service uses the IGN BD TOPO® database.

The API documentation is available \href{https://geoservices.ign.fr/documentation/services/services-geoplateforme/geocodage}{here}.
}
\examples{
\dontrun{
# Get accepted values for POI categories (Geocode API) ----

gs_get_poi_categories()
} 
}
