% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/rgeoservices-package.R
\docType{package}
\name{rgeoservices-package}
\alias{rgeoservices}
\alias{rgeoservices-package}
\title{rgeoservices: A R Client for the IGN Geoservices API}
\description{
A set of functions to query the API provided by the French National Institute of Geographic and Forest Information (IGN) through its online Geoplateform \url{https://geoservices.ign.fr/documentation/services/services-geoplateforme}.
}
\seealso{
Useful links:
\itemize{
  \item \url{https://github.com/ahasverus/rgeoservices}
  \item Report bugs at \url{https://github.com/ahasverus/rgeoservices/issues}
}

}
\author{
\strong{Maintainer}: Nicolas Casajus \email{rdev.nc@gmail.com} (\href{https://orcid.org/0000-0002-5537-5294}{ORCID}) [copyright holder]

}
\keyword{internal}
